Modified

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args){
-        long price = 4599;
-        long bonus = price / 20;
+        int price = 4599;
+        int bonus = price / 20;
         System.out.println(bonus);
     }
 }
